feat(#27): pipeline CI — clippy -D warnings + cargo test + rustfmt
- Añade .github/workflows/ci.yml con tres jobs: clippy, test, fmt
- Corrige todos los avisos de clippy detectados en el workspace:
  redundant closures, needless_range_loop, type_complexity,
  unnecessary_cast, dead_code, while-let loop, is_multiple_of
- Aplica cargo fmt --all para que el check de formato pase desde el
  primer commit

Closes #27

+++ crates/bitturbulence-client/src/daemon/filler.rs
@@ -16,8 +16,8 @@ use super::KEEPALIVE_INTERVAL;
 // ── Bucle del filler (conexión entrante) ──────────────────────────────────────
 
 pub async fn run_peer_filler(
-    conn:     &PeerConnection,
-    ctx:      &Arc<FlowCtx>,
+    conn: &PeerConnection,
+    ctx: &Arc<FlowCtx>,
     _peer_id: &[u8; 32],
 ) -> Result<()> {
     // Stream 1 ya fue gestionado por handle_inbound (Hello / HelloAck).
@@ -68,7 +68,7 @@ pub async fn run_peer_filler(
                     Ok((fi, pi)) => {
                         let _ = ctrl_w.send(Message::HavePiece {
                             file_index: fi as u16,
-                            piece_index: pi as u32,
+                            piece_index: pi,
                         }).await;
                     }
                     Err(tokio::sync::broadcast::error::RecvError::Lagged(n)) => {
